+++ src/main/kotlin/wordle/io/Printer.kt
@@ -1,22 +1,38 @@
-package wordle.view
+package wordle.io
 
+/**
+ * 값을 출력해주는 객체입니다.
+ */
 object Printer {
 
+    // TODO: 6번을 외부로부터 입력 받는 형식으로 수정필요
+    /**
+     * Wordle 게임 시작을 안내하는 인삿말을 출력합니다.
+     */
     fun introduce() {
         println("WORDLE을 6번 만에 맞춰 보세요.\n" +
                 "시도의 결과는 타일의 색 변화로 나타납니다.")
     }
 
+    /**
+     * 값 입력을 요구하는 메세지를 출력합니다.
+     */
     fun requestInput() {
         println("정답을 입력해 주세요.")
     }
 
+    /**
+     * tile을 출력합니다.
+     */
     fun viewTile(tile: String) {
         println()
         println(tile)
         println()
     }
 
+    /**
+     * 최종 결과를 출력합니다.
+     */
     fun result(count: Int, currentCount: Int, tile: String) {
         println("${currentCount}/${count}")
         viewTile(tile)
